requirements.txt
pyTelegramBotAPI>=4.21.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langgraph-checkpoint-redis>=0.0.1
redis[asyncio]>=5.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
